Move ISR subscription field in fix module
This allow to have a common module for both l10n_ch_payment_slip and l10n_ch_qriban
