+++ l10n_ch_fix_isr_reference/models/invoice.py
@@ -2,12 +2,73 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 import re
 
-from odoo import api, models
+from odoo import api, fields, models
 from odoo.tools import mod10r
+from odoo.addons.l10n_ch_base_bank.models.bank import pretty_l10n_ch_postal
 
 
 l10n_ch_ISR_NUMBER_LENGTH = 27
 l10n_ch_ISR_ID_NUM_LENGTH = 6
+
+
+class FutureAccountInvoice(models.Model):
+    """Rewrite field l10n_ch_isr_subscription to get the right field"""
+    _inherit = 'account.invoice'  # pylint:disable=R7980
+
+    l10n_ch_isr_subscription = fields.Char(
+        compute='_compute_l10n_ch_isr_subscription',
+        help=(
+            "ISR subscription number identifying your company or your bank "
+            " to generate ISR."
+        ),
+    )
+    l10n_ch_isr_subscription_formatted = fields.Char(
+        compute='_compute_l10n_ch_isr_subscription',
+        help=(
+            "ISR subscription number your company or your bank, formated"
+            " with '-' and without the padding zeros, to generate ISR"
+            " report."
+        ),
+    )
+
+    @api.depends(
+        'partner_bank_id.l10n_ch_isr_subscription_eur',
+        'partner_bank_id.l10n_ch_isr_subscription_chf')
+    def _compute_l10n_ch_isr_subscription(self):
+        """ Computes the ISR subscription identifying your company or the bank
+        that allows to generate ISR. And formats it accordingly
+
+        """
+
+        def _format_isr_subscription_scanline(isr_subscription):
+            # format the isr for scanline
+            isr_subscription = isr_subscription.replace('-', '')
+            return (isr_subscription[:2]
+                    + isr_subscription[2:-1].rjust(6, '0')
+                    + isr_subscription[-1:])
+
+        for record in self:
+            isr_subs = False
+            isr_subs_formatted = False
+            if record.partner_bank_id:
+                bank_acc = record.partner_bank_id
+                if record.currency_id.name == 'EUR':
+                    isr_subscription = bank_acc.l10n_ch_isr_subscription_eur
+                elif record.currency_id.name == 'CHF':
+                    isr_subscription = bank_acc.l10n_ch_isr_subscription_chf
+                else:
+                    # we don't format if in another currency as EUR or CHF
+                    isr_subscription = False
+
+                if isr_subscription:
+                    isr_subs = _format_isr_subscription_scanline(
+                        isr_subscription
+                    )
+                    isr_subs_formatted = pretty_l10n_ch_postal(
+                        isr_subscription
+                    )
+            record.l10n_ch_isr_subscription = isr_subs
+            record.l10n_ch_isr_subscription_formatted = isr_subs_formatted
 
 
 class AccountInvoice(models.Model):
